implement load, store. memory api uses u32

+++ src/memory.rs
@@ -1,8 +1,8 @@
 pub trait Addressable {
-    fn read(&self, addr: u16) -> Option<u8>;
-    fn write(&mut self, addr: u16, value: u8) -> bool;
+    fn read(&self, addr: u32) -> Option<u8>;
+    fn write(&mut self, addr: u32, value: u8) -> bool;
 
-    fn read2(&self, addr: u16) -> Option<u16> {
+    fn read2(&self, addr: u32) -> Option<u16> {
         if let Some(x0) = self.read(addr) {
             if let Some(x1) = self.read(addr + 1) {
                 return Some((x0 as u16) | ((x1 as u16) << 8));
@@ -11,16 +11,16 @@ pub trait Addressable {
         None
     }
 
-    fn write2(&mut self, addr: u16, value: u16) -> bool {
+    fn write2(&mut self, addr: u32, value: u16) -> bool {
         let lower = value & 0xff;
         let upper = (value & 0xff00) >> 8;
         self.write(addr, lower as u8) && self.write(addr + 1, upper as u8)
     }
 
-    fn copy(&mut self, from: u16, to: u16, n: usize) -> bool {
+    fn copy(&mut self, from: u32, to: u32, n: usize) -> bool {
         for i in 0..n {
-            if let Some(x) = self.read(from + (i as u16)) {
-                if !self.write(to + (i as u16), x) {
+            if let Some(x) = self.read(from + (i as u32)) {
+                if !self.write(to + (i as u32), x) {
                     return false;
                 }
             } else {
@@ -30,9 +30,9 @@ pub trait Addressable {
         true
     }
 
-    fn load_from_vec(&mut self, from: &[u8], addr: u16) -> bool {
+    fn load_from_vec(&mut self, from: &[u8], addr: u32) -> bool {
         for (i, b) in from.iter().enumerate() {
-            if !self.write(addr + (i as u16), *b) {
+            if !self.write(addr + (i as u32), *b) {
                 return false;
             }
         }
@@ -55,7 +55,7 @@ impl LinearMemory {
 }
 
 impl Addressable for LinearMemory {
-    fn read(&self, addr: u16) -> Option<u8> {
+    fn read(&self, addr: u32) -> Option<u8> {
         if (addr as usize) < self.size {
             Some(self.bytes[addr as usize])
         } else {
@@ -63,7 +63,7 @@ impl Addressable for LinearMemory {
         }
     }
 
-    fn write(&mut self, addr: u16, value: u8) -> bool {
+    fn write(&mut self, addr: u32, value: u8) -> bool {
         if (addr as usize) < self.size {
             self.bytes[addr as usize] = value;
             true
